requirements/requirements.txt
ipython
pandas
pyarrow
numpy
pyspark
requests
python-dotenv
minio
apache-airflow[postgres]
apache-airflow-providers-apache-spark
streamlit
s3fs
pendulum